+++ Source/CIM.RemoteManager.App/CIM.RemoteManager.Core/Models/Sensor.cs
@@ -9,11 +9,11 @@
         public string SerialNumber { get; set; }
         public string Name { get; set; }
         public string SensorType { get; set; }
-        public decimal Scale { get; set; }
-        public decimal Offset { get; set; }
+        public double Scale { get; set; }
+        public double Offset { get; set; }
         public int TimeStamp { get; set; }
-        public decimal AverageValue { get; set; }
-        public decimal CurrentValue { get; set; }
+        public double AverageValue { get; set; }
+        public double CurrentValue { get; set; }
         public int DisplayConversionCode { get; set; }
         public int DecimalLocation { get; set; }
         public string StatisticsTotalCalcSettings { get; set; }
